[#172351208] Handle abort in adding new payment method (#1806)
* abort the insertion of a new payment method during the verification

* refactoring Checkout3DsComponent > Checkout3DsModal to support the hardware back in modal

* clean code

* clean code

* replace ActionSheet.show -> Alert.alert
